requirements.txt
asgiref==3.7.2
Django>=4.0,<4.1
django-filter==23.5
djangorestframework==3.14.0
importlib-metadata==6.7.0
Markdown==3.4.4
Pygments==2.17.2
pytz==2018.9
sqlparse==0.4.4
typing_extensions==4.7.1
zipp==3.15.0
